#193 Refatorando pastas e locais de teste
