+++ hortum_mobile/test/services/codec_string_test.dart
@@ -2,17 +2,14 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:hortum_mobile/services/codec_string.dart';
 
 main() {
-  String str = 'Teste';
-  String strCoded = 'dGVzdGU=';
-
   test('Testing the encode of a string to base64 string', () {
-    encodeString(str);
+    String result = encodeString('Teste');
 
-    expect(strCoded, 'dGVzdGU=');
+    expect(result, 'VGVzdGU=');
   });
   test('Testing the decode of a string base64 to string', () {
-    decodeString(strCoded);
+    String result = decodeString('VGVzdGU=');
 
-    expect(str, 'Teste');
+    expect(result, 'Teste');
   });
 }
