+++ hortum_mobile/test/views/home_productor/home_productor_page_test.dart
@@ -46,30 +46,4 @@ main() {
 
     expect(find.byKey(Key('removeAnnoun')), findsOneWidget);
   });
-
-  test('Testing the method manipulateData on ProdLoggedAnnounDataApi', () {
-    ProdLoggedAnnounDataApi announProd = ProdLoggedAnnounDataApi();
-    dynamic element = {
-      "username": "Usuário Teste",
-      "email": "usuário@gmail.com",
-      "idPicture": null,
-      "announcements": [
-        {
-          "username": "Usuário Teste",
-          "idPictureProductor": null,
-          "name": "Folha Verde",
-          "type_of_product": "Alface",
-          "description": "Alface plantado na fazenda",
-          "price": 5.0,
-          "idPicture": null
-        }
-      ]
-    };
-    announProd.announcements = element['announcements'];
-
-    announProd.manipulateData();
-
-    expect(announProd.announcements[0]['price'], "R\$ 5,00");
-    expect(announProd.announcements[0]['username'], "Usuário");
-  });
 }
