+++ hortum_mobile/test/views/profile/profile_form_test.dart
@@ -2,7 +2,6 @@ import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hortum_mobile/components/form_field.dart';
-import 'package:hortum_mobile/data/update_user_backend.dart';
 import 'package:hortum_mobile/globals.dart';
 import 'package:hortum_mobile/views/profile/components/profile_form.dart';
 import 'package:mockito/mockito.dart';
@@ -14,22 +13,6 @@ main() {
   final dio = DioMock();
   TestWidgetsFlutterBinding.ensureInitialized();
   var url = 'http://$ip:8000/users/update/';
-
-  group('Testing class UpdateUserAPI:', () {
-    UpdateUserAPI updateUser = UpdateUserAPI(dio);
-    actualUser.tokenAccess = 'token';
-    String responseMatcher = 'Dados alterados';
-    test('Update User Successful', () async {
-      when(dio.patch(url, data: anyNamed('data'), options: anyNamed('options')))
-          .thenAnswer((_) async => Response(
-              data: jsonEncode(responseMatcher),
-              requestOptions: null,
-              statusCode: 200));
-      int responseActual = await updateUser.updateUser(email: '', username: '');
-
-      expect(responseActual, 200);
-    });
-  });
 
   group('Testing profile_form:', () {
     Widget makeTestable() {
